Review de la totalite des instructions

+++ src/op/zjmp.c
@@ -6,7 +6,7 @@
 /*   By: gguichar <gguichar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/17 02:17:41 by gguichar          #+#    #+#             */
-/*   Updated: 2019/02/20 22:57:47 by gguichar         ###   ########.fr       */
+/*   Updated: 2019/02/21 01:39:53 by gguichar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,9 +31,8 @@ int			zjmp(t_env *env, t_process *cur_process, uint8_t *bytes)
 	fill_decode(env, cur_process, &decode, 1);
 	ret = decode_args(&decode, bytes + 1, DIR_CODE << 6, SHORT_DIR) + 1;
 	if (cur_process->carry == 1)
-		ret = ((int)decode.tab[0].value % IDX_MOD);
+		ret = (int)decode.tab[0].value % IDX_MOD;
 	if (env->debug == DEBUG_ON)
-		debug_mode((int)decode.tab[0].value % IDX_MOD
-				, (int)cur_process->carry);
+		debug_mode((int)decode.tab[0].value % IDX_MOD, cur_process->carry);
 	return (ret);
 }
